Refactoring and test for validator
